refactor: update winner page for using index import
