+++ src/features/car/generate-cars/generate-cars.ts
