+++ src/features/car/generate-cars/lib/random-color/random-color.test.ts
