+++ src/features/car/generate-cars/model/types.ts
