+++ src/pages/winners/model/init.ts
