+++ src/widgets/index.ts
@@ -1,3 +1,5 @@
+export { winnerTableHeader } from './winner-table-header/ui/winner-table-header';
+export { winnerTable } from './winner-table/ui/winner-table';
 export { carList } from './car-list/ui/car-list';
 export { carPanel } from './car-create-panel/ui/car-create-panel';
 export { createPagination } from './pagination/ui/pagination';
