+++ vite.config.js
@@ -13,7 +13,12 @@ export default defineConfig({
       enabled: false,
       reporter: ['text'],
       include: ['src/**/*.{js,ts}'],
-      exclude: ['src/**/*.test.{js,ts}', 'src/**/*.d.ts', 'src/**/*types.ts'],
+      exclude: [
+        'src/**/*.test.{js,ts}',
+        'src/**/*.d.ts',
+        'src/**/*types.ts',
+        'src/shared/test/mock.ts',
+      ],
       thresholds: {
         global: {
           statements: 70,
